xkb: rename xkbgeom.h to xkbgeom_priv.h
Our naming convention for headers is that private ones (not exported,
ie. not part of xf86 sdk) shall be suffixed by "_priv.h".
