+++ xkb/xkbfmisc.c
@@ -44,7 +44,7 @@
 #include "dix.h"
 #include "xkbstr.h"
 #include <xkbsrv.h>
-#include "xkbgeom.h"
+#include "xkbgeom_priv.h"
 
 unsigned
 _XkbKSCheckCase(KeySym ks)
